Rearanged files and refactored the pipeline of seeding script
