+++ package.json
@@ -9,7 +9,7 @@
     "react-dev": "webpack --mode development --watch",
     "build": "webpack --mode production",
     "server-dev": "nodemon server/start.js",
-    "seed": "node seedDatabase/index.js"
+    "seed": "node SDC/seed.js"
   },
   "repository": {
     "type": "git",
@@ -48,12 +48,15 @@
     "nodemon": "^2.0.7",
     "path": "^0.12.7",
     "pexels": "^1.2.1",
+    "pg": "^8.6.0",
+    "pg-hstore": "^2.3.4",
     "prompt": "^1.1.0",
     "qs": "^6.10.1",
     "react": "^16.14.0",
     "react-dom": "^16.14.0",
     "react-dom-server": "0.0.5",
     "react-router-dom": "^5.2.0",
+    "sequelize": "^6.6.5",
     "sinon": "^10.0.0",
     "styled-components": "^5.3.0",
     "supertest": "^6.1.3",
@@ -72,6 +75,7 @@
     "css-loader": "^5.1.3",
     "csv-write-stream": "^2.0.0",
     "csv-writer": "^1.6.0",
+    "dotenv": "^10.0.0",
     "enzyme": "^3.11.0",
     "enzyme-adapter-react-16": "^1.15.6",
     "enzyme-to-json": "^3.6.2",
@@ -84,7 +88,6 @@
     "jest": "^26.6.3",
     "lorem-ipsum": "^2.0.3",
     "mocha": "^9.0.1",
-    "pg": "^8.6.0",
     "pg-copy-streams": "^5.1.1",
     "source-map-loader": "^2.0.1",
     "style-loader": "^2.0.0",
